This is part of the process of switching my singleton entities to use plists rather than sql

+++ SHModels/SHModels/SHModelTools.h
@@ -8,4 +8,4 @@
 
 @import Foundation;
 
-uint shCalculateLvl(uint lvl,uint range);
+NSUInteger shCalculateLvl(NSUInteger lvl,NSUInteger range);
